MAINT: move diddynamic from dev to public package, add top-level exports
