+++ moderndid/diddynamic/__init__.py
@@ -1,6 +1,6 @@
 """Dynamic covariate balancing for Difference-in-Differences."""
 
-import moderndid.dev.diddynamic.format as _format
+import moderndid.diddynamic.format as _format
 
 from .container import DynBalancingHetResult, DynBalancingHistoryResult, DynBalancingResult
 from .dyn_balancing import dyn_balancing
