+++ moderndid/diddynamic/estimation/coefficients.py
